+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 未发布的更新
 
+## 0.7.17
+
 ### 其他
 
 改用 [`pdm`](https://github.com/pdm-project/pdm) 管理项目依赖.